exec in start_p1.sh overlays the script1.sh onto start_p1.sh
This results in proper management by procmon of the secon script started
by start_p1.sh. Else if we kill the script started by start_p1.sh, all
is well as start_p1.sh did not start in backhround and is blocking. As
soon as sscript1.sh is killed/terminated, start_p1.sh also terminates
and procmon handles this.

But if we kill directly start_p1.sh then procmon doesnt track the
script1.sh started by that. So it started again everything but we have
an orphan script1.sh from previous run and that gets owned by init
process and never gets terminate till we epxlicitely kill it.

Starting with exec in start_p1.sh overlays the image of script1.sh on
start_p1.sh and it taked the pid of start_p1.sh which is already managed
by procmon and all works well.

Signed-off-by: Shobhit Kumar <kumar@shobhit.info>