CTRL+C Mystry solved. SIGINT goes to all foreground processes
SIGINT when raised using CTRL+C goes to all foreground processes on that
shell, including all nested programs. So CTRL+C on procmon shell which
has launhed app1 (in foreground) resulted SIGINT also being sent to app1
which was not handling it and so default handler was killing it before
SIGTERM could reach it.

Register SIGINT in app1 to ignore SIGINT and now SIGTERM is working
well.

Signed-off-by: Shobhit Kumar <kumar@shobhit.info>