Create a background daemon application and see if we can monitor that
NO, we cannot as it does a fork() inside itself to get new process as a
daemon closing the parent process. So no way to track that PID.

Added just as an example. If we enable this commit we will get
application starting quickly error and after 5 retries procmon will
abort. Though this will end up in having 6 dameons which will then be
inherited by init process as parent.

Signed-off-by: Shobhit Kumar <kumar@shobhit.info>